requirements.txt
botocore==1.19.4
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
gunicorn==20.0.4
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
jsonschema==3.2.0
Markdown==3.2.1
MarkupSafe==1.1.1
oauthlib==3.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
tornado==6.0.4
tweepy==3.9.0
urllib3==1.25.9
userpath==1.4.1
whitenoise==5.2.0
yarn==0.0.8
google-cloud-language==2.0.0
google-cloud==0.34.0